requirements.txt
openai==1.30.1
langchain==0.1.17
setuptools~=69.2.0
PyYAML~=6.0.1
build==1.2.1